P727-037: Use informational popup to notify that search restarted
A new Informational_Popup package has been created, allowing to
display ephemeral informational popups to be displayed, using icons
to inform the user on a particular event. This package use Gtk
transitions (via the Gtk_Revealer widget) to show/hide informational
popups.

By default, we now display a informational popup with a loopback icon
to notify the user when the serach has restarted from the beginning of
the current file.

Change-Id: Id1e7db27daef5a52df58cb2aaa1527f813c43c9c