Fix various warnings when compiling the doc
The package gi.repository cannot be imported on our servers, so we make
 it optional for the doc

Replace some properties with a @property.
This doesn't impact the rendering in HTML, but encourages to properly
document the type of the property and makes it easier to search in the
file for all properties

Make sure that we display values for enumeration types.
 This was broken since sphinx now imports gps_utils, and the definition
 of enum() from gps_utils is used.
Make sure that enumeration names are used in parameter lists, rather than
 the equivalent value. We have to display them as strings, but it is still
 better to see "GPS.Browsers.Style.Arrow.NONE" than 0

Part of P930-013

Change-Id: I698006d142c48ab90729931df912109e0def2654