P727-037: Navigate in matched occurences stack when pressing 'Backspace'
When the Search view is in incremental mode, pressing the backspace key
in the search entry allows the user to navigate among the previous
matched occurrences.

Refactoring has been made in order to support this capability: instead
of giving factory functions to create a particular search context we now
give a Search_Module_Type, which has a primitive Create_Context for this
particular purpose. In addition, the search module owns a stack of
search occurrences, allowing the Search view to push/pop/highlight a given
one.

Change-Id: I2522745ca47b771d4447091eea8247a232230435