Support the "show hidden files" preference in Files view
This preference now applies to both files and directories.
It also applies to multiple views (Project, Files and VCS explorer).

Change the way we fill the Files view, so that we insert a larger number
of files at each idle, rather than a single one (we spend up to 40ms
to do so). This means the file view is filled almost instantaneously
on fast machines, and does not impact the performance on slower disks.

PA21-030

Change-Id: I18c66f0ecb72c20af8920b95cb250a13d21345ef