PA17-045: Fix the 'Edit->Paste Previous' capability
Pastes on Gtk_Text_Buffer widgets can be asynchronous: in this
case, trying to retrieve the text buffer cursor's position right
after a call to Gtk.Text_Buffer.Paste_Clipboard will not return the
cursor's position after the paste.

To avoid this problem, we now connect to the 'paste-done' signal
to retrieve the cursor's position after a paste, since this signal
is emitted every time a paste has been done.

This signal was disabled in Src_Editor_Buffer due to an old bug
in Gtk but this bug does not appear anymore.

Change-Id: I51ebfab1f33c6d0bf25b80bd7385b1ef7164e993