Run git ls-tree and git status in parallel
This means we can get the status of all files about twice as fast as
before. We also systematically run "ls-tree" now, instead of caching
the result. This makes sure that when a new file is saved, its status
is properly computed as well.

(join): new function to wait for two or more promises to complete

Removing misleading trace in src_editor that seems to show Save_To_File
was called twice.

O116-004

Change-Id: I872741ad6f084bfe3a9dd22e839f83d21485c344