Add support for a callback when Ensure_Status_For_* have completed
This is needed for the Commits view, since it needs to know all files
under VCS control, even those that will not be added to the cache because
they have the VCS's default status (for instance untracked files in the
case of CVS and subversion). This information is only available in the
cache after Ensure_Status_For_* has finished executing, possibly in the
background depending on the python implementation.

O116-004

Change-Id: I37e6edf689ee009b466b6a8a0cba0cfcea31c816