Fix scrolling when pasting with multiple views
Change d60997b0a0463aa9a2d7cfc02a09b190b7ccd0ce introduced a
regression. If we have multiple views of the same buffer, then
pasting in any of the view will scroll all of them (the default
behavior of gtk+, which we were overriding).
That commit was also adding one callback on paste-done every time
we did a paste on a buffer, resulting in extra work

PA17-045

Change-Id: I4558d687fbfaed4b38ebea7e43a5a588dfca4060