(GPS.EditorBuffer.get): fix handling of force=True
We were not forcing the reload. As a result, autognatpp.py sometimes
seemed to have no effect because the file was not reloaded properly
from the disk.

Partial rewrite of the gnatpp plugin: we now use python rather than
the GPS shell.
(autognatpp.py): pass the name of the file to gnatpp, rather than
 rely on the context. Fixes cases where this plug-in did not work
 as expected.

PB09-037

Change-Id: I6b86f4abd853f8f883ae055cc24b29b62a6ec1fe