PA04-010: Add the possibility to register file templates using aliases
The Language_Handlers.Assistants package defines a new FileTemplate
Python class that allows users to register file templates
using aliases.

When registering a file template, an action and a contextual menu
are created for it, allowing the user to create a new file with the
right naming scheme and prefill it with the expanded text of its
associated alias.

Change-Id: I9a72ecec4f7f149f48c2d5f43f4078b909e031a9