Fix too much scrolling when expanding node in Project view
Now that we temporarily detach and then reattach the model from the view,
calling Scroll_To_Cell no longer works as expected. gtk+ does not properly
compute the visible range, and thus always moves the expanded node to the
top of the view, even if it was already visible.

(Add_Row_Children): make sure the expanded status is preserved, so remove
 the dummy node last. This removes the need to call Expand_Row again while
 in the expand row callback.

Change-Id: I20e155168f3824cc4c94ebe23d2a23f162663c1c