PB23-010 : Use the GNATCOLL.Projects.Artifacts_Dir function in GPS
A new IDE'Artifacts_Dir has been created in GNATCOLL, which can be specified
with a directory path in which all the files generated by GPS for a given
project will be put (xref database files, locations).

This attribute defaults to the project's object directory if not specified.

The documentation has been updated accordingly.

Change-Id: I133bbfab224903b9db78327079396bd021d15725