P727-037: Seach view incremental mode enhancements
A new 'incremental search' action has been created to open the Search
view in incremental mode even if the hidden preference is not set. This
action is now binded to 'control + s' in the Emacs mode key bindings,
instead of the older 'isearch'.

Moreover, executing the action that opens the Search view when its entry
has already the focus (and if in incremental mode) searches for the
first/next occurrence, like Emacs does.

Change-Id: I9c0c71e6c5f120acc2defb2e49eb315617d8885c