Change the python API to read the history of a VCS repository
This is now more efficient. Filling the history view is also faster.
For the gps repository (40070 lines in the full history), it used to
take 1min50s. "git log" itself takes about 2s.
This is now down to 1min10, most of which is just setting the gtk+
tree model.
The new implementation also uses an idle on the Ada side, so GPS
is never blocked while filling the History view. We used to block
when the tree was filled for each line read by the git plugin.

O116-004

Change-Id: I63c3040e1a50767855fa3ea5365344df2805d8e2