Implement filter in the History view
We delegate all the filtering work to the VCS engine, for efficiency:
otherwise, we would have to always query all commits, to have all
information in memory. This is slow on large repositories like the
linux kernel (up to 1min to get all commits, their full commit message
and the list of modified files), and needs a lot of memory on the GPS
side.
Instead, we let the VCS engine do the filtering, so that we can take
advantage of it (for instance limiting the number of commits returned
to 2000, which significantly speeds up the parsing, or searching in the
diff in the case of git).

The history view now works for large repositories like the linux kernel
(5s to display the first 2000 commits and all the branches).

O116-004

Change-Id: I073c33d4fbd9fab5f0b6d9ab7c1c7ab8685067ed