PB16-048: Display a preferences assistant for newcomers
A preferences assistant is now displayedif no GPS_HOME was found when
starting GPS.

The purpose of this assistant is so set a number of critical preferences
(e.g: color theme, keymaps etc.) before starting GPS for the first time.

The actual pages that this assistant will contain should be discussed: the
current ones are there for testing purposes.

Change-Id: I15f2170626fb49841f611158ef618a16768e21a8