PC21-006: Avoid referencing Preferences in Preference_Groups
Instead of storing directly the Preference ojbjects in the
Preference_Group objects, we now only store the preferences names that
belong to this group and, when needed, retrieve the actual Preference
object from the Preference_Manager map.

This is needed in order to allow preferences to be displayed among
multiple pages (e.g: one in the Preferences Editor dialog and one in
the Preferences Assistant dialog).

Change-Id: I7f97a9b2fec94b46e7b67c2628174366186e6093