PB16-048: Display the new welcome dialog
Use the new GPS welcome dialog in GPS.Main instead of the old one.

This dialog displays all the recently used projects in its left pane.

In its right pane, it displays a header and the different actions
that can be done: 'Create new project', which then displays the project
templates' assistant, and 'Open project', which allows the user to
open any project.

Work to be done:

  . Add a GPS logo in the header
  . Create the 'Checkout project from VCS' option

Change-Id: I847d51e5b095a4b043a5fe499bdb5be50130db44