Further optimize parsing VCS status
For git, we now reuse the output of "git ls-tree", until the user
manually forces a reload. This command is used to find the list of
files managed by git, but this list doesn't change afterwards without
also impacting "git status".

Do not create two File_Sets after parsing the status of files. Such sets
might be slow to create for qgen (94_000 files, 0.4s to create each
set). A single such set is needed, and only when calling the hook.

Reserve capacity for the VCS status cache, to speed things up (640ms
down to 400ms for the initial fill, on qgen)

Properly take into account the default status of files for the VCS
engine, when deciding whether to execute the hook. We were missing a
dispatching call, and as a result we had to create a file_sets for
the 94_000 files, then execute the vcs_file_status_change_hook. Instead,
we have nothing to do here.

O116-004

Change-Id: If89d4d24cdc20b0c830be523ff72884ab1ac5c49