Remove xref computations rendez-vous points in the editor
The editor is now responsible for requesting an asynchronous
refresh of the semantic_tree, in a timeout after it has been
edited.

Remove bottlenecks that were happening when requesting this
information for the needs of blocks.

Remove fields that are related to the constructs from the
editor, these are no longer needed.

For PC20-002.

Change-Id: I24ba050d2f7e6b724c253d0386ec3b1d518fa417