Q106-021 (1/2): Move the project dependencies editor in Project Properties
A new page has been introduced in the Project Properties editor dialog
allowing users to edit the project dependencies directly from here.

This new page allows users to drag a known project (i.e: a project
present in the whole project tree or known from the project paths) and
drop it in a 'Dependencies' tree view to create a new dependency.

Change-Id: I4a3b4ef6a5c6d8ba56bad98e8e6efbd3cd20dd7d