Improve filtering in History view
When a filter is set, we should not display the graph, since too many
commits are missing and the graph would not make sense.
In the case of git, support special prefixes in the filter to clarify
the intent: "author:" and "code:" to filter on one or the other, rather
than on the message.

O116-004

Change-Id: I9662921f304503203e86373a20f76b027c526dd6