Let VCS plugins what to do when double-clicking
In particular, double-clicking on a stash will now apply it to the
working directory.
Plug-ins can also enhance the tooltips displayed in the branches view.
Add support for creating and deleting branches in git.
Add support for creating and deleting tags in git.
Add support for stashing changes, or dropping existing stash in git.
Add support for deleting remote branches.
Add support for creating and removing cvs tags.

O116-004

Change-Id: Ie01c4840086a1ab598be1ec8718ad0a7d09a63a6