Fix computing history for a single file
We were systematically compute the whole history, then clearing the view
and finally applying the filter.
A new parameter has been added to Get_Or_Create_View, so that we can
disable the automatic initial refresh (for the full view), when we
intend to apply a filter afterwards.

Fix incorrect reset of the layout information when multiple "log"
commands have been queued. We were reseting before queuing, so when
the second log was finally executed we were starting with pre-set
data, which resulted in a Constraint_Error when computing the layout.

Q117-025

Change-Id: Ib7d92a500b6cf4d3a075c8d82088667cb9f46dea