Make search functionality independent of the vsearch widget
Up to now, we were storing data in the vsearch widget itself,
and then counting on this widget never being destroyed, but getting
hidden instead.
This has proven fragile, and requires inelegant code in the generic_view
package.

This patch changes this invariant, and makes sure the search
functionalities are available even when the dialog is not present.
Preparatory work for removing the inelegant/dangerous code in
generic_view.

For Q118-015.

Change-Id: I8f3a413501d0121bd5048b7a56c43e90659fd125