Remove dangerous code relying on emitting delete events
Remove the concept of views that have the "Hide_Rather_Than_Close"
property (the vsearch was the only view which needed this, and this
is no longer the case).

This removes the possibility for client views to intercept and react to
a "delete-event" on their window.

This allows us to completely rely on the delete_event on a view
in order to know when to memorize the size of this view.

In turn, this allows us to remove dangerous code consisting in
sending delete_events on the toplevel window and letting this
window react in order to store its size: this code was dangerous
because the deletion of a toplevel window is asynchronous, and
it could occur that the MDI listed the child as being up even
after the window was being deleted. With the change, rather than
emitting delete on toplevel windows, we can use the MDI functions
to close children, and therefore the MDI is no longer in a danger
of thinking that some windows are available when they are in
fact being deleted.

For Q118-015.

Change-Id: I4b92f353f617e82d3857b12debbc7ea7eb201b1c