Fix issue with incremental search
Do not consider that incremental search is on if the view is
not present.

Merge On_Pattern_Combo_Changed into On_Pattern_Combo_Changed_After,
to make sure that the pattern is updated before acting on a change
in the contents.

Add a flag to control the behavior of interactive search, making sure
we are not initiating interactive searching until the user has entered
something in the search dialog.

Fixes Q127-016

Change-Id: Id7c1326808984f37ae0cb29f66f720a7da946ef8