PC12-019: Revamp of the memory usage view
The memory usage view has been revamped.

It is now able to display memory usage information about the
different memory regions and sections of an executable, even
when the linking phase has failed.

This information is displayed in a tree view, with the memory
regions at the first level and the associated memory sections
as children, if any.

Change-Id: Iace86607a5e51af79e7f881ff517210152c890a7