Q215-038: Use the welcome dialog as transient window
We now pass the welcome dialog to all the welcome dialog actions
callbacks so that they can set it as transient window when
they spawn a dialog (e.g: when clicking on 'Create new project').

This ensures that the dialogs spawned from the welcome dialog are
always put on top of: the Set_Keep_Above procedure was not
reliable (depends too much on the windows manager).


Change-Id: I2ddd5e0544aa5696884b0451ead7e1f08efbfe57