Q215-029 Support for relocating codepeer DB
(Inspection): New parameter DB_Directory to override the DB directory
stored in Inspection_Info.xml which may not be the right one. This
parameter will be ignored by old versions of gps_codepeer_bridge (with
the default being to use the DB stored in Inspection_Info.xml), and taken
into account by the new gps_codepeer_bridge (changes already made in
CodePeer as part this ticket).

Change-Id: I91f645c876872ed391bbf95fd76d5c56fad5a27c