Do not store a Kernel in a Entity...
Or in an entity database. This is not a great abstraction, since it
does not deal well with the case of No_Entity: should this constant
have a Kernel? Or should absolutely all the clients test for the
validity of the entity before using Entity.Kernel?

Instead, use the kernel that's stored in the Clang_Module_Id.

First step for investigating the regression in libclang_xref_cpp,
for Q221-011.

Change-Id: I9998eb9f156526d680dbc05a92555aaa547afe41