Q313-006: Fix the 'From Project %p' entry in the Search view
The 'From Project %p' entry was not displayed anymore in the Search
view's scope selector (i.e: the combo box after the 'In' label).

Moreover, the behavior of the Search view regarding this entry has been
fixed: we only want to refresh it when the project view has changed, not
every time the search is being reset.

This avoids having this entry automatically set to the root's project
instead of the project found in the context when creating/opening the
Search view.

Change-Id: Ia1761922d74bc2b84ee297c5411999042811be7f