GNATdoc: Disable assertion on subtypes of generic formals
This patch disables an assertion on subtypes of generic formals
since the Xref decoration of subtypes of generic formals is
incomplete. It avoids crash processing the GPS sources.

In addition, when checking the decoration of the tree nodes,
this patch coverts errors into warnings since in most cases
they are not serious errors and the backend can generate
the html output.

For PC02-022


Change-Id: Ifd25b3bc57a956b21bab42da9d7d13393e33347f