Q217-010: Avoid creating secondary messages for same line in generics
When dealing with generic units, several messages with different
weights (e.g: info vs medium) can potentially point to the same
location.

In this case, we want them to be separated in the Locations view
so that the message with the higher weight takes precedence over
the ones with a lower weight.

Change-Id: I3293441d219e62b4c44df7f2fd171aa8ef16cc8c