Q510-037: Modify default key shortcuts
An explicit key shortcut has been added for "select all" and the
key shortcuts for jumping to backward/forward locations and
word based navigation/selection have been modified.

Add additional key shortcuits for "goto beginning/end of line" on Mac
OS (primary-Right and primary-Left are standard in additional to control-A
and control-E).

Change-Id: Ie29c3b6f7643764688c652a02f2b2bb0a24c8ef8