Q728-010 Using location of the vc for prove check command
This commit makes use of new information provided by gnatprove which now
contains the precise location of the VC check (the location displayed in
messages is not enough). This should solve a problem where the limit-line
information provided to gnatprove needed to be manually changed to allow
execution of a manual prover.
This commits is associated to a commit in spark that generates said infos.

* share/plug-ins/spark2014.py
(build_msg_full_text): Removing this function from the class for it to be
used in on_prove_check function.
(get_comp_text): From a message, returns the text that is used as a key
(in msg_id) and/or displayed to the user.
(on_exit): Adding a global map 'map_msg' from message text to associated
locations (3 fields: check_file, check_line, check_col).
(act_on_extra_info): Fills map_msg with appropriate information from
*.spark file.
(limit_line_option): Changed to use new line and col.
(on_prove_check): Using new line and col.

Change-Id: Ic220ba0003e4a9ceec20c4853b5dd5f6956d567d