Q706-011: Use file as index when editing switches
When editing switches for a file, we should make sure that the
file's base name is used as the index of the associated attribute
(either Default_Switches or Switches).

Also, make sure to register the Switches attribute when no explicit
attribute has been speicified when registering a tool: GPS should
be able to understand both attributes in that case.

Change-Id: Ia2e17401c0400e8d791d11c00d2f07f4f96f8e1c