Q823-022: Help text box in build targets dialog
Add the <command-help> tag in the XML description of the targets.
Get the help message and show it in the build targets dialog.
The help message defined in a <target> replaces the help message
defined in a <target-model>.
The help is shown in a text view with a monospace font.

Change-Id: I1887bbaae39dc5646a61064e563c3c70872ce958