Q926-009: Improve Project Properties behavior with scenarios
When applying changes made in Project Properties to all the possible
values of a given scenario variable, we should make sure that no
useless switch case is written in the resulting .gpr file.

This is ensured by the fact that we pass All_Scenarios when calling
GNATCOLL.Project.Set_Attribute in that case, instead of calling it
mutilple times with a different scenario value.

The scenario selector is now expanded by default, to make it more visible
in the Project Properties.

Support has been added in testing utility files to get switches listed
in the currently selected Project Properties page and to select
scenarios in the scenario selector.

Change-Id: I80559439ca479d4a0c746c9f7f0c1ebc9c6d9215