Q929-008 gnat_server not in PATH
When installing spark, gnat_server is not put into the PATH because users
are not expected to execute it directly. So, to launch it from GPS, we use
its path relatively to gnatprove. This also checks that gnat_server
actually exists: if it does not, an error message is printed and use of
GPS is resumed.

* share/plug-ins/spark2014.py
(print_error): Used to print errors on the console 'Messages'.

(compute_gnatserver_path): compute the path of gnat_server using the
path of gnatprove.

(start_ITP): computation of the gnat_server path is now done at global
level. If gnat_server was not detected (is_itp), nothing is done and an
error is printed.

(exit_ITP): Condition on is_itp to send an error if gnat_server is not
detected.

Change-Id: I1c5ce2036da0098bcc865e276a966c439f3b0b96