QA02-025 Change function used to get the current compunit
get_compunit_for_message can now parse a message containing
counterexamples. Also, now uses this function in on_prove_itp.

* share/plug-ins/spark2014.py
(reg[1-3]): Changed to be used with re.search instead of match. Removed "$"
at the end of the regexp because counterexamples can be in this line of
text.

(get_compunit_for_message): Make this function available at library level
to use it in on_prove_itp. Change re.match to re.search: the results should
be the same, we just need a smaller regexp.

Change-Id: I44da9cda0f56411f9b94d518c81ef098f1334310