Q926-021: Fix focus issues with omnisearch from floating views
When performing a search from a floating view, we should make sure that
the focus is given to the omnisearch entry: that's why we need to call
Present on the GPS main window in that case.

Make sure to give the focus back to the floating editor when leaving the
omnisearch entry: again, we should call Present on the floating editor's
window for that.

Change-Id: Ia6df7b1cba0e792e1723c4c61cae2bdfce667464