QA23-047 Rework the implementation of the theme selection
Add a high level object for describing a Theme and a Color.
Add a level of indirection in order to centralize all the
references to names of GPS preferences in one place, and
simplify the code of theme importers and theme definitions.

Remove the mechanism for having screenshots that represent
themes, use Gtk Labels instead, and generate them on the
fly.

And support for Gruvbox themes in the theming engine.

Change-Id: I2042bc7f0c72de16042f126dd9b3dbcb9328887b