QA16-017: Fix Locate in Files View
The whole project is not loaded in cache and by default the
loading of a new file is asynchrone. So add an attribute to
be synchrone when searching for a file in the view.
Non fixed issue: the scrolling doesn't work the first time for
a big project, the searched file is selected but too many files
are added so the view scroll back to the top (ex: GTKAda project)

Change-Id: I5de0a7c871b1142e9b0dcedb4c3520b0cfcf3f9c