Contextual menu improvements
Internal data is organized now as tree instead of a flat list.

This allows to add filter on entire submenu in order to increase
performance by reducing calling of filters on each element of submenu.

It can be done for example for CVS submenu which can be deactivated
when a project doesn't use version control at all.

The same for Build & Run submenus which can be active only when we
build contextual menu for file or project in the Project view
and so on.

This patch is the first step for introducing tree-like list of
contextual menu items. Filters will be added as a second step.

Change-Id: I5a5d95b378bda8b8a9fdf2fe3e67052fe42640e2