Q831-016: Implement the initial Learn module mechanism
A Learn package has been created. It can be used to register providers that
can create learn items (Gtk_Flow_Box_Child derived widgets) that will be
displayed in the Learn view.

These learn items can be filtered out by the Learn view when necessary to
display only the learn items that are relevant in the current context.

Change-Id: I54b40946066a0ce0fc8c8ac77aa62b437913c9ab