QB12-002 Switch directory for Coqide
For Coqide 8.5pl2 (also for 8.5pl3 and 8.6), for unknown reasons, we need
to execute coqide from the directory of the .v file. This commit is an
ad hoc modification that implements switching directories.

* share/plug-ins/spark2014.py
(act_on_extra_info): launching coqide or any manual prover (paths are
absolute and it should not be a problem) from the right directory.

Change-Id: I57e9aca94bc598f3b106a0c11784c93678a23a98