QB15-005 Running manual proof process with option directory
This commit corrects a bug triggered by the changing of current working
directory in spark plugin. Now, the process that required this change is
called with option directory of GPS.Process.

* share/plug-ins/spark2014.py
Adding a directory as argument of tree.start for launching gnat_server
process. Removing the change of directory.

* share/plug-ins/spark2014/itp_lib.py
Adding directory option when launching GPS.Process.

Change-Id: Ic4b46a650b703cad0b3dcd7fd50bd2d5ef98bb9c