QB20-015: Use a proper search context in GPS.Kernel.Search.Actions
Instead of relying on current context to show only the valid actions,
we now rely on a separate search context that is explicitly set in
Gtkada.Entry_Completion.

This allows to perform the search with the context that was set just
before entering the search field, but without restoring it, which
avoids having conflicts with editor actions while being in the
omnisearch (e.g: backspace not sent to editor).

Change-Id: I4829f7de4b27be881c4f0272a6760b469dea4bec