QB07-035 Adding a menu entry for debugging proof session
This adds a menu entry that allows to show the whole session for a file.
It is useful to update existing manual proof or to show a proof.

* share/plug-ins/spark2014.py
(debug_session): Removed and added as local parameter to functions
(prove_check_context): Adding a local debug_session. It allows to run itp
even when no message is selected.
(start_ITP): Adding a debug argument.
(on_prove_itp): Adding an optional debug argument. If the debug flag is
set, the arguments generated for launching itp_server (coming from the
message) cannot be found. So, only the file is specified in this case.

* share/plug-ins/spark2014/gnatprove_menus.xml
Adapting calls to prove_check_context and on_prove_itp. Adding a menu,
"Edit Manual Proof" which is called with debug flags set to true.

Change-Id: If2ee84d804db19b372b084d5c224be943a08374c