Q831-016: Display more actions in the Learn view
Many actions have been added to the Learn view.

The Learn view mechanism has been improved: we not connect on the
On_Context_Changed hook to refilter the learn items (needed for
some of the associated actions).

A 'for-learning' attribute has been added for the XML-defined targets
in order to be able display them in the Learn view too. 

Change-Id: I63d1605579811977b5c6015c24e86443312d14ef