QC15-016 Fix 'Show local changes for X in editor'
on Windows. We reset read-only permissions before rewriting the temporary
file. On closing vdiff we close temporary file even if still exists, due
to windows prevents deletion of read-only file.
Also we change order of opening 'real' and 'ref' file, so that after
closing vdiff 'real' file will not be in a split state.

Change-Id: Ieeb91daa12f0b3a6226340b13d0145419de214d1