R106-001: GPS hang when loading coverage data
Exiting GPS when the Coverage Report is filled will provoc
a large dump of XML in Code_Analysis_Tree: a node for each file
covered in the report. Furthermore, when the project is modified,
GPS will try to refresh the Coverage Report using the node XML
Code_Analysis_Tree. So when this node contains thousands of children
GPS need a long time to finish.

This patch prevents GPS from dumping redundant and useless data
in perspectives6.xml. And prevents GPS from analysing gcov data
when using gnatcoverage.

Change-Id: If748a4a116447846a8fe8490a478dd91716aae4a