R109-015 Itp: add debug options
Add a debug option to debug calls to gnat_server with --debug-stack-trace.

* share/plug-ins/spark2014/itp_lib.py
Add variable debug_file which holds the location of the file to be used
for the error output of the server.

* share/plug-ins/spark2014.py
Change launching of the server when a file is provided in debug_file. It
outputs stack_trace to this file.

Change-Id: I46e7764a0cb48721ce7d351a9f8ae2e1b88fc7e7