R116-057: Fix incremental search when pressing backspace
When some text is matched, push the search occurrence only when different
from the last one that was pushed.

Also, don't use the search occurrences' stack mechanism when the search
has failed but use the normal backspace behavior instead: we only want to
go to a previous occurrence (if any) when the text matches.

Change-Id: I9a7ef1c6dbf3047921124e07771bbdf2b8b75444