Q116-042 fix implementation of the "Python Library" help
This action relies on pydoc, so only display the menu if
pydoc is found on the PATH. i

This menu was also duplicated, remove the "old" way of
displaying it, preferring the entry defined in menus.xml.

Also make this menu work regardless of the context,
it makes no sense to have it work only when a Python file
is selected.

Change-Id: Ibaa7aa14554a5afb170fdebcae6534e1b427fb5d