R206-042 Fix memory corruption in the Outline view
The outline view was holding a reference to a Semantic_Tree. This is
not valid: these objects may be destroyed in the background by the
engine that produces them.

The stored Semantic_Tree was used to be able to find a node that
corresponds to a given location - instead, we now do this by storing
the slocs for nodes in the outline tree model.

As part of this, make the Outline view completely asynchronous,
only refreshing itself when it received an up-to-date tree.

Change-Id: I62a22d5d88c3c226e0db75a257a6267753b1dc6f