Q209-036: add a button for running GNATtest executable in GNATemulator
This button will change depending if the current test harness is a test
list or a single test. In the case of the test list, we do not output
the messages in the Locations view.

These buttons are showed when a gnatemu is present on the PATH and when
the target architecture is not the same as the one of the host.

Change-Id: I3c9c14225ec80a0b998c81fbe0700e9fd11b2caa