R131-028: Add a fallback when CSS theme colors are not defined
A fallback has been added to the themes engine when the CSS preference
that exposes the theme's colors is not defined: in that case, we try
now to deduce the user's theme from the editor's style preference and
apply it again to set this CSS preference.

When no theme can be deduced, we only expose the CSS editor's colors
to make sure that the GPS base colors (which are now deduced from
the editor's ones) are not broken.

Change-Id: I4d8765c1c1525a43509dd54ec4baaaed7dfa5256