R314-022: fix bad regexp for connection tool
The regexp used to detect that the connection tool was connected to the
board was not working in all cases. Specifically, when using OpenOCD to
connect to a Atmel board. It should now work with all boards, we check
that OpenOCD outputs the number of hardware breakpoints on the
board.

Change-Id: I7ffcd88754e7e10089d9acf80a7f05d58044cdbd