R131-036 Adjust computation of gutter foreground
Use a mix between the editor foreground and background to make the line
numbers slightly less visible than the text.

Fix the calculation for mixing colors.

Adjust the theme preview window to the fact that the gutter background
is now the same as the editor background.

Adjust the color for ephemeral highlighting to be more readable in
the majority of cases.


Change-Id: I0ce080ee41603c4663b7bd14e3eb6c5fafccb932