Q927-021: Allow empty connection tool for baremetal workflows
Some users spawn board connection tools outside of GPS because
we don't support them: in that case, we still want them to use
the 'Debug on Board' toolbar button, which is more practical and
intuitive than having to click on 'Debug/Initialize'.

Moreover, the 'Build & Run' and 'Build & Debug' workflow buttons are
only meant for native development, so we should hide them when the
project is non-native.

The doc has also been updated.

Change-Id: Icc7008be8b5ff6e4a42789cd9604c096f830c476