R404-027: Fix line numbers display
The way we calculate the width needed to draw the line numbers has
been fixed: we now calculate if from the maximum number of line
characters (e.g: 4 for the line 4523) and the width needed by
one line character. Space for 2 line characters is reserved if
the files contains less than 10 lines.

We now draw line numbers from the right margin of the space
allocated to all line numbers, instead of centering them.

The additional margins that are scattered all around have been
properly named and modified when needed.

The 'Line numbers width' preference has been renamed to 'Gutter right
margin' and now controls the space added between the whole gutter and
the editor's text.

Change-Id: If5b9b4d4075f02f29f7a3e5245d66880693a509e