R403-023: Don't accept invalid values for typed scenario variables
When clicking on the Scenario's view apply button, GPS now checks if
the user has entered invalid values.

If it's the case, the project view is not recomputed and an error
infobar is displayed at the top. 'Wrong' icons are also displayed
next to the invalid values, making it easier to indentify which
scenario variables  are incorrect.

Change-Id: Ie47acf04dfc5a787f3be9044b1185f242b5f4674