R411-004: Add visual feedback on the omnisearch
A spinner is now displayed while loading the search results: when a
sufficient number of search results have been inserted, the tree view
displaying the search results is shown and the spinner is hidden.

A label displaying "No results" is displayed when no search results have
been found for the given text.

The popup's height has been reduced by two.

Change-Id: I6c9252047d8186c04f625010b7ba6b68e0a7a89e