R405-041: Correct the behavior of the special lines
When trying to get a location in a special lines GPS
was returning a location in the first editable line, thus
creating a lot of unwanted behaviors. GPS now returns 0
and each actions need to properly handle this case.

For the python API adds the function Ensure_Valid_Position:
a wrapper around Is_Valid_Position which raises an exception
if the position is invalid. The exception is then treated
in the python API.

Other issue:
Keep the style of the special lines when folding/unfolding.
Also load the Data of the saved lines when Unhiding a line.


Change-Id: I386112ca7a375bab7651262b328b9f427ed6697b