R406-022: Add "continue to line" clickable icon on editors
When a debugger is running and if the debugger was started, a
"Continue to line" clickable icon is displayed on the editor's
current line.

Clicking in it has for effect to continue the xecution until the given
line is reached.

A new Debugger.Continue_Until_Location method has been added in order to
support that without using temporary breakpoints (except for LLDB).

The debugger.py plugin has been removed: it was declaring a custom
"continue till line" action that is not needed anymore.

Change-Id: I22f2a96db4738be7a431bdb82aa37b435b161915