R411-042 fix detection of file changed on disk
We check the timestamp of files on disk every time we switch
the focus, and we were supposed to do this every time we save
as well, but this was broken: the implementation of Check_Monitored_Files
was silently doing a reload in this case.

Fix the return type of Check_Monitored_Files according to the
specification, clarify the specification for the non-interactive
use of this function, and add a parameter which allows the function
to monitor only one given file.

This new parameter allows us to use this function to display
the dialog in the specific case of one editor being saved, rather
than having to implement a different dialog there.

Also add a trace which allows us to test the checking of timestamps.

Change-Id: I0e7f149820c16d6f360b164ff140c0f216d396be