Q522-019: Add a preference to always show the gnathub rule
If the preference is disabled, then a rule will only be visible
if one of its severities is selected in the filter views.
If the preference is enabled, then a rule will always be visible
and it will show the current amount of visible messages and the
total amount of messages.

Change-Id: I4a85bede531b8ef6ba107977cd55c70a9c28f444