R131-030: Trust the database before the semantic tree
For the entities messages, trust gnathub and create the entity
accordingly.
For the files messages, check in the semantic tree if the message
is located in an entity. If true, then add this entity in the
Analysis tree.
The semantic tree and the gnathub database often doesn't have the
same values for the entities' columns so don't use the columns when
mangling the entities' names.

Change-Id: Ifbbcd64e29a5cdf2b4bd6287227747e54d5c0844