R702-006: Add an interface to run python code during the template creation
We now call the python script multiple times during the template
instantiation. The python script is expected to provide an object with
the two following methods:

- get_next_pages : returns a list of widget used as successive gtk
  assistant pages, after the project is selected.

- on_apply : called at the end of the template instanciation,
  the current working directory is where the template has been
  installed.

We can't add an attribute to the Project_Template record because it
would make project-templates dependent upong Gtk components. The
solution is to create a record containing a Project_Template and its
associated Script_Object. We then store this record directly in each
widget representing a template page.

Change-Id: I876e0db4812d9d5eb3ff3bdc4ac373bf284d5aec